fix: fix conponent name
